requirements.txt
streamlit
numpy
pandas
opencv-python
Pillow
tensorflow
scikit-learn
matplotlib
plotly
fpdf